requirements.txt
streamlit
openai
requests
fastapi
uvicorn[standard]
PyJWT